fix(rubocop): re-enable BlockLength and ModuleLength
Split remaining hotspots (instrumentation events, optimized M2M helpers,
measure dataset filters) and enable both cops with documented excludes for
engines route DSL, nested-set association blocks, rake namespaces, CDS
synchronizer, and migrations.

Issue: No ticket/issue
