+++ app/models/measure.rb
@@ -218,230 +218,10 @@ class Measure < Sequel::Model
     result.compact
   end
 
-  module DatasetFilters
-    def with_additional_code_sid(additional_code_sid)
-      return self if additional_code_sid.blank?
-
-      where(additional_code_sid:)
-    end
-
-    def with_additional_code_type(additional_code_type_id)
-      return self if additional_code_type_id.blank?
-
-      where(additional_code_type_id:)
-    end
-
-    def with_additional_code_id(additional_code_id)
-      return self if additional_code_id.blank?
-
-      where(additional_code_id:)
-    end
-
-    def join_measure_conditions
-      association_right_join(:measure_conditions)
-        .where(Sequel.~(measures__measure_sid: nil))
-    end
-
-    def with_certificate_type_code(type)
-      return self if type.blank?
-
-      where(measure_conditions__certificate_type_code: type)
-    end
-
-    def with_certificate_code(code)
-      return self if code.blank?
-
-      where(measure_conditions__certificate_code: code)
-    end
-
-    def with_certificate_types_and_codes(certificate_types_and_codes)
-      combine_pair_conditions(
-        certificate_types_and_codes,
-        :measure_conditions__certificate_type_code,
-        :measure_conditions__certificate_code,
-      )
-    end
-
-    def join_footnotes
-      association_right_join(:footnotes)
-        .exclude(measures__measure_sid: nil)
-    end
-
-    def with_footnote_type_id(footnote_type_id)
-      return self if footnote_type_id.blank?
-
-      where(footnotes__footnote_type_id: footnote_type_id)
-    end
-
-    def with_footnote_id(footnote_id)
-      return self if footnote_id.blank?
-
-      where(footnotes__footnote_id: footnote_id)
-    end
-
-    def with_footnote_types_and_ids(footnote_types_and_ids)
-      combine_pair_conditions(
-        footnote_types_and_ids,
-        :footnotes__footnote_type_id,
-        :footnotes__footnote_id,
-      )
-    end
-
-    def with_measure_type(condition_measure_type)
-      where(measures__measure_type_id: condition_measure_type.to_s)
-    end
-
-    def valid_to(last_effective_timestamp)
-      where('measures.validity_start_date <= ?', last_effective_timestamp)
-    end
-
-    def terminated
-      where('measures.validity_end_date IS NOT NULL')
-    end
-
-    def with_geographical_area(area)
-      where(geographical_area_id: area)
-    end
-
-    def effective_start_date_column
-      Sequel.function(:coalesce,
-                      :measures__validity_start_date,
-                      :base_regulation__validity_start_date,
-                      :modification_regulation__validity_start_date)
-    end
-
-    def effective_end_date_column
-      Sequel.function(:coalesce,
-                      :measures__validity_end_date,
-                      :base_regulation__effective_end_date,
-                      :base_regulation__validity_end_date,
-                      :modification_regulation__effective_end_date,
-                      :modification_regulation__validity_end_date)
-    end
-
-    def with_generating_regulation
-      association_left_join(:base_regulation, :modification_regulation)
-        .where do |_query|
-          (Sequel.qualify(:base_regulation, :base_regulation_id) !~ nil) |
-            (Sequel.qualify(:modification_regulation, :modification_regulation_id) !~ nil)
-        end
-    end
-
-    def with_regulation_dates_query_non_current
-      with_generating_regulation
-        .select_append(Sequel.as(effective_start_date_column, :effective_start_date))
-        .select_append(Sequel.as(effective_end_date_column, :effective_end_date))
-    end
-
-    def with_regulation_dates_query
-      with_generating_regulation
-        .select_append(Sequel.as(effective_start_date_column, :effective_start_date))
-        .select_append(Sequel.as(effective_end_date_column, :effective_end_date))
-        .where do |_query|
-          if model.point_in_time
-            start_date = effective_start_date_column
-            end_date   = effective_end_date_column
-
-            (start_date <= model.point_in_time) &
-              ((end_date >= model.point_in_time) | (end_date =~ nil))
-          else
-            true # .where method needs _something_ to AND to the query
-          end
-        end
-    end
-
-    def with_seasonal_measures(measure_type_ids, geographical_area_ids)
-      start_of_range = Time.zone.today.beginning_of_year
-      end_of_range = Time.zone.today.end_of_year + 1.year
-
-      select(
-        :measure_sid,
-        :goods_nomenclature_item_id,
-        :measure_type_id,
-        :geographical_area_id,
-        :validity_start_date,
-        :validity_end_date,
-      )
-        .where(measure_type_id: measure_type_ids)
-        .where(geographical_area_id: geographical_area_ids)
-        .exclude(validity_end_date: nil)
-        .where('validity_start_date >= ?', start_of_range)
-        .where('validity_end_date <= ?', end_of_range)
-        .where(Sequel.lit('(validity_end_date::date - validity_start_date::date) NOT IN ?', ANNUAL_DURATION_DAYS))
-    end
-
-    def dedupe_similar
-      # Needs with_regulation_dates_query and only works within time machine but should be used before not after
-      select(Sequel.expr(:measures).*)
-        .distinct(:measure_generating_regulation_id,
-                  :measure_generating_regulation_role,
-                  :measure_type_id,
-                  :goods_nomenclature_sid,
-                  :geographical_area_id,
-                  :geographical_area_sid,
-                  :additional_code_type_id,
-                  :additional_code_id,
-                  :ordernumber)
-        .order(*DEDUPE_SORT_ORDER)
-    end
-
-    def without_excluded_types
-      exclude(measures__measure_type_id: MeasureType.excluded_measure_types)
-    end
-
-    def overview
-      where do
-        overview_types = [
-          { measures__measure_type_id: MeasureType::SUPPLEMENTARY_TYPES },
-          {
-            measures__measure_type_id: MeasureType::THIRD_COUNTRY,
-            measures__geographical_area_id: GeographicalArea::ERGA_OMNES_ID,
-          },
-        ]
-
-        overview_types << if TradeTariffBackend.uk?
-                            {
-                              measures__measure_type_id: MeasureType::VAT_TYPES,
-                              measures__geographical_area_id: GeographicalArea::AREAS_SUBJECT_TO_VAT_OR_EXCISE_ID,
-                            }
-                          else
-
-                            {
-                              measures__measure_type_id: MeasureType::VAT_TYPES,
-                              measures__geographical_area_id: GeographicalArea::ERGA_OMNES_ID,
-                            }
-
-                          end
-
-        Sequel.|(*overview_types)
-      end
-    end
-
-    def excluding_licensed_quotas
-      exclusion_criteria = Sequel.|(
-        *QuotaOrderNumber::LICENSED_QUOTA_PREFIXES.map do |prefix|
-          Sequel.like(:ordernumber, "#{prefix}%")
-        end,
-      )
-
-      exclude(exclusion_criteria)
-    end
-
-  private
-
-    # Builds an OR-combined Sequel condition from a collection of [col_a, col_b]
-    # value pairs and applies it as a WHERE clause. Returns the dataset unchanged
-    # if the pairs collection is empty.
-    def combine_pair_conditions(pairs, col_a, col_b)
-      return self if pairs.none?
-
-      conditions = pairs.map { |a, b| Sequel.expr(col_a => a) & Sequel.expr(col_b => b) }
-      where(conditions.reduce(:|))
-    end
-  end
-
   dataset_module do
-    include DatasetFilters
+    include DatasetFilters::AssociationFilters
+    include DatasetFilters::ScopeFilters
+    include DatasetFilters::OverviewFilters
   end
 
   def_column_accessor :effective_end_date, :effective_start_date
